+++ raphtory/src/db/graph/views/mod.rs
@@ -1,6 +1,6 @@
+pub mod cached_view;
 pub mod deletion_graph;
 pub mod layer_graph;
-pub mod masked_graph;
 pub mod node_subgraph;
 pub mod node_type_filtered_subgraph;
 pub mod property_filter;
